requirements.txt
djangorestframework==3.12.4
django-cors-headers==3.7.0
pytz==2021.3